+++ Data/Config.cs
@@ -1,12 +1,14 @@
 ﻿using System.Collections.Generic;
 using AutoMapPins.Common;
 using UnityEngine;
+
 // ReSharper disable ConvertToConstant.Global
 // ReSharper disable FieldCanBeMadeReadOnly.Global
 // ReSharper disable CollectionNeverUpdated.Global
 // ReSharper disable UnassignedField.Global
 // ReSharper disable ClassNeverInstantiated.Global
 // ReSharper disable UnusedMember.Global
+// ReSharper disable MemberCanBePrivate.Global
 
 namespace AutoMapPins.Data;
 
@@ -17,7 +19,7 @@ public abstract class Config
         public Dictionary<string, Pin>? IndividualConfiguredObjects = null!;
         public List<string>? CategoryConfiguredObjects = null!;
     }
-    
+
     public class Pin : SerializedToStrings
     {
         public string? Name = null!;
@@ -31,14 +33,38 @@ public abstract class Config
 
     public class PinColor : SerializedToStrings
     {
-        public float Red = 1f;
-        public float Green = 1f;
-        public float Blue = 1f;
-        public float Alpha = 1f;
+        internal static PinColor White = new() { Red = RGBMax, Green = RGBMax, Blue = RGBMax, Alpha = RGBMax };
+        
+        private const int RGBMin = 0;
+        private const int RGBMax = 255;
+        private const float Divider = (float)RGBMax;
+
+        public int Red;
+        public int Green;
+        public int Blue;
+        public int Alpha;
+
+        internal PinColor ClampColor()
+        {
+            Red = Clamp(Red, RGBMin, RGBMax);
+            Green = Clamp(Green, RGBMin, RGBMax);
+            Blue = Clamp(Blue, RGBMin, RGBMax);
+            Alpha = Clamp(Alpha, RGBMin, RGBMax);
+            return this;
+        }
 
         public Color FromConfig()
         {
-            return new Color(Red, Green, Blue, Alpha);
+            float red = Clamp(Red, RGBMin, RGBMax) / Divider;
+            float green = Clamp(Green, RGBMin, RGBMax) / Divider;
+            float blue = Clamp(Blue, RGBMin, RGBMax) / Divider;
+            float alpha = Clamp(Alpha, RGBMin, RGBMax) / Divider;
+            return new Color(red, green, blue, alpha);
+        }
+
+        private static int Clamp(int value, int min, int max)
+        {
+            return value < min ? min : value > max ? max : value;
         }
     }